requirements.txt
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-cloud-pubsub==2.29.0
httplib2==0.22.0
python-dotenv==1.1.0
requests==2.32.3
certifi==2025.4.26
psycopg2==2.9.10
sqlalchemy==2.0.40
